+++ keat_scripts/release.sh
@@ -27,7 +27,7 @@ if [ "$input" != "" ]; then
 fi
 
 poetry run keats bump $VERSION
-
+VERSION=$(poetry run keats v)
 ################################
 # Setup
 ################################
@@ -50,7 +50,7 @@ fi
 
 printf "$CINPUT Would you like to publish this package to a repo?$END\n"
 printf "$CINPUT New repos can be configures using $CINFO 'poetry config repositories.<reponame> <url>' $END\n"
-printf "$CINPUT Repository name (default: pypi): $END"
+printf "$CINPUT Repository name (example: pypi; ENTER to skip): $END"
 read input
 if [ "$input" != "" ]; then
     REPO=$input
@@ -114,6 +114,7 @@ printf "\n$SEP Publishing $SEP\n"
 
 
 if [ "$REPO" != "" ]; then
+    poetry config repositories.$REPO
     printf "$CWARN Are you sure you want to publish $NAME $VERSION to $REPO ([y]/n)?: $END"
     read input
     if [ "$input" == "n" ]; then